# Speed up norm_backward