--- a/tools/autograd/gen_variable_type.py
+++ b/tools/autograd/gen_variable_type.py
@@ -690,7 +690,7 @@ def load_aten_declarations(path):
 
         declaration['base_name'] = declaration['name']
 
-        # if the return value is missing a name, call it 'output'
+        # if the return value is missing a name, call it 'result'
         for ret in declaration['returns']:
             if 'name' not in ret:
                 assert len(declaration['returns']) == 1