Also avoids a CUDA synchronization in norm_backward.

Here's a benchmarking script. **You must comment out norm in variable.py to see any difference.**

```python
import torch; from torch.autograd import Variable; v = Variable(torch.randn(1024,1024), requires_grad=True); vc = Variable(v.data.cuda(), requires_grad=True); g = torch.randn(1024); gc = torch.randn(1024).cuda()

%timeit v.norm().backward()
%timeit v.norm().backward()
%timeit v.norm(0).backward(g)
%timeit v.norm(0).backward(g)

%timeit vc.norm().backward()
%timeit vc.norm().backward()
%timeit vc.norm(0).backward(gc)
%timeit vc.norm(0).backward(gc)
```

Before:
```
100 loops, best of 3: 3.96 ms per loop
100 loops, best of 3: 4.04 ms per loop
100 loops, best of 3: 8 ms per loop
100 loops, best of 3: 7.94 ms per loop
1000 loops, best of 3: 544 µs per loop
1000 loops, best of 3: 532 µs per loop
1000 loops, best of 3: 619 µs per loop
1000 loops, best of 3: 620 µs per loop
```

After:
```
100 loops, best of 3: 2.99 ms per loop
100 loops, best of 3: 2.97 ms per loop
100 loops, best of 3: 6.02 ms per loop
100 loops, best of 3: 5.66 ms per loop
1000 loops, best of 3: 369 µs per loop
1000 loops, best of 3: 342 µs per loop
1000 loops, best of 3: 417 µs per loop
1000 loops, best of 3: 421 µs per loop
```