@ezyang 2017-10-20
I didn't verify the math, but it looks good to me :)